+++ packages/core/src/rules/common/spec-ref-validation.ts
@@ -2,12 +2,13 @@ import { Oas2Rule, Oas3Rule } from '../../visitors';
 import { isRef } from '../../ref-utils';
 
 export const SpecRefValidation: Oas3Rule | Oas2Rule = () => {
-  const nodesToSkip = [
+  const nodesToValidate = [
     'Schema',
     'Response',
     'Parameter',
     'RequestBody',
     'Example',
+    'Examples',
     'Header',
     'SecurityScheme',
     'Link',
@@ -17,9 +18,7 @@ export const SpecRefValidation: Oas3Rule | Oas2Rule = () => {
 
   return {
     any(_node, { report, rawNode, rawLocation, type }) {
-      const shouldCheck = !nodesToSkip.includes(type.name);
-
-      if (shouldCheck && isRef(rawNode)) {
+      if (!nodesToValidate.includes(type.name) && isRef(rawNode)) {
         report({
           message: 'Field $ref is not expected here.',
           location: rawLocation.child('$ref').key(),
