--- packages/core/src/rules/oas3/spec-ref-validation.ts
@@ -1,29 +1,0 @@
-import { Oas3Rule } from '../../visitors';
-import { isRef } from '../../ref-utils';
-
-export const SpecRefValidation: Oas3Rule = () => {
-  const nodesToValidate = [
-    'Schema',
-    'Response',
-    'Parameter',
-    'RequestBody',
-    'Example',
-    'Examples',
-    'Header',
-    'SecurityScheme',
-    'Link',
-    'Callback',
-    'PathItem',
-  ];
-
-  return {
-    any(_node, { report, rawNode, rawLocation, type }) {
-      if (!nodesToValidate.includes(type.name) && isRef(rawNode)) {
-        report({
-          message: 'Field $ref is not expected here.',
-          location: rawLocation.child('$ref').key(),
-        });
-      }
-    },
-  };
-};
